# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "paths": {
        "/quiz": {
            "post": {
                "operationId": "0e4662952fcff0a8127e2aad5106b95d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Le nom du quiz à créer",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz créé avec succès"
                    },
                    "400": {
                        "description": "Le nom du quiz existe déjà"
                    },
                    "500": {
                        "description": "Erreur lors de la création du quiz"
                    }
                }
            },
            "delete": {
                "operationId": "9a41dbaaa82f23c106757189510397c3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz supprimé avec succès"
                    },
                    "400": {
                        "description": "Le nom du quiz n'existe pas"
                    }
                }
            }
        },
        "/correct": {
            "post": {
                "tags": [
                    "corrections"
                ],
                "operationId": "43f32f8abec07d9c061d089f662a678f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz corrigé"
                    },
                    "304": {
                        "description": "Erreur lors de la préparation de la correction ou lors de la correction"
                    }
                }
            }
        },
        "/omr_errors_resolved": {
            "post": {
                "tags": [
                    "corrections"
                ],
                "operationId": "6a12348c4020caf6ba0e360ec85772f6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execution du script OMR errors réussie sur le quizz"
                    },
                    "304": {
                        "description": "Erreur lors du script OMR errors"
                    }
                }
            }
        },
        "/upload_copies": {
            "post": {
                "tags": [
                    "corrections"
                ],
                "operationId": "5ad399648d6d25f97c6c5c741d0cb7b4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "file",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier de correction uploadé avec succès"
                    },
                    "500": {
                        "description": "Erreur lors de l'upload du fichier de correction"
                    }
                }
            }
        },
        "/upload_correction": {
            "post": {
                "tags": [
                    "corrections"
                ],
                "summary": "Upload de la correction du quiz",
                "operationId": "uploadCorrection",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nom du quiz",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fichier de correction à uploader",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Fichier à uploader",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fichier de correction uploadé avec succès"
                    },
                    "500": {
                        "description": "Erreur lors de l'upload du fichier de correction"
                    }
                }
            }
        },
        "/upload_list": {
            "post": {
                "tags": [
                    "lists"
                ],
                "summary": "Upload de la liste des élèves du quiz",
                "operationId": "uploadList",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nom du quiz",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fichier de liste à uploader",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Fichier à uploader",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fichier de liste uploadé avec succès"
                    },
                    "500": {
                        "description": "Erreur lors de l'upload du fichier de liste"
                    }
                }
            }
        },
        "/check_correction": {
            "get": {
                "operationId": "ef574a1eafa7550924568e1967179422",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des quiz"
                    },
                    "500": {
                        "description": "Erreur lors de la récupération de la liste des quiz"
                    }
                }
            }
        }
    }
}